Handle missing talent records and add error boundary
