+++ talentify-next-frontend/app/api/messages/route.ts
@@ -44,8 +44,8 @@ export async function POST(req: NextRequest) {
     .select()
     .single()
 
-  if (error) {
-    return NextResponse.json({ error: error.message }, { status: 500 })
+  if (error || !data) {
+    return NextResponse.json({ error: error?.message ?? 'not found' }, { status: 500 })
   }
 
   return NextResponse.json(data, { status: 201 })
