+++ talentify-next-frontend/app/api/schedule/route.ts
@@ -44,8 +44,11 @@ export async function POST(req: NextRequest) {
     .select()
     .single()
 
-  if (error) {
-    return new Response(JSON.stringify({ error: error.message }), { status: 500 })
+  if (error || !data) {
+    return new Response(
+      JSON.stringify({ error: error?.message ?? 'not found' }),
+      { status: 500 }
+    )
   }
 
   return new Response(JSON.stringify(data), { status: 201 })
